feature: several changes
