+++ core-methods/src/main/java/com/mercadopago/sdk/android/coremethods/data/remote/request/InstallmentsRequest.kt
@@ -3,7 +3,6 @@ package com.mercadopago.sdk.android.coremethods.data.remote.request
 import java.math.BigDecimal
 
 internal data class InstallmentsRequest(
-    val productId: String? = null,
     val bin: Int? = null,
     val processingMode: String? = null,
     val amount: BigDecimal? = null,
