+++ core-methods/src/main/java/com/mercadopago/sdk/android/coremethods/data/remote/request/PaymentMethodsRequest.kt
@@ -1,6 +1,5 @@
 package com.mercadopago.sdk.android.coremethods.data.remote.request
 
 internal data class PaymentMethodsRequest(
-    val productId: String? = null,
     val bin: Int? = null,
 )
