+++ core-methods/src/main/java/com/mercadopago/sdk/android/coremethods/domain/model/params/GetPaymentMethodsParams.kt
@@ -1,6 +1,5 @@
 package com.mercadopago.sdk.android.coremethods.domain.model.params
 
 internal data class GetPaymentMethodsParams(
-    val productId: String? = null,
     val bin: Int? = null,
 )
