+++ core-methods/src/main/java/com/mercadopago/sdk/android/coremethods/domain/usecase/GetInstallmentsUseCase.kt
@@ -17,7 +17,6 @@ internal class GetInstallmentsUseCase(
     ): Result<List<Installment>, ResultError> {
         return repository.getInstallment(
             GetInstallmentParams(
-                productId = "",
                 bin = bin.toIntOrNull(),
                 amount = amount,
                 processingMode = processingMode,
